requirements.txt
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
black==25.1.0
boltons==21.0.0
bracex==2.6
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-option-group==0.5.7
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
defusedxml==0.7.1
Deprecated==1.2.18
Django==5.2.1
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-ipware==7.0.1
django-model-utils==5.0.0
django-ratelimit==4.1.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
exceptiongroup==1.2.2
face==24.0.0
glom==22.1.0
googleapis-common-protos==1.70.0
icalendar==6.3.1
idna==3.10
importlib_metadata==7.1.0
inflection==0.5.1
isort==6.0.1
jsonfield==3.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
Markdown==3.8
markdown-it-py==3.0.0
mdurl==0.1.2
mypy_extensions==1.1.0
mysqlclient==2.2.7
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
packaging==25.0
pathspec==0.12.1
pbr==6.1.1
peewee==3.18.2
pillow==11.2.1
pipdeptree==2.28.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==4.25.8
pydotplus==2.0.2
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.2.3
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-ipware==3.0.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.3
rich==13.5.3
rpds-py==0.24.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stevedore==5.4.1
swapper==1.4.0
tomli==2.0.2
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcmatch==8.5.2
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.23.0
